requirements.txt
Flask==3.1.0
google-generativeai==0.8.6
gspread==6.2.1
oauth2client==4.1.3
requests==2.32.3
python-dotenv==1.2.1
gunicorn==23.0.0